코드 개선 및 서비스 구현 추가
