+++ sharedPrompts/src/main/java/org/example/sharedprompts/global/jwt/SecurityConfig.java
@@ -24,7 +24,7 @@ public class SecurityConfig {
     private final TokenRedisService tokenRedisService;
     private final OAuth2SuccessHandler oAuth2SuccessHandler;
     private final OAuth2FailureHandler oAuth2FailureHandler;
-    private final CustomOAuth2UserService customOAuth2UserService;
+    private final CustomOAuth2UserServiceImpl customOAuth2UserService;
     private final JwtAuthenticationEntryPoint jwtAuthenticationEntryPoint;
     @Bean
     public PasswordEncoder passwordEncoder() {
